+++ src/main/kotlin/me/daegyeo/netflixchecker/api/controller/AuthController.kt
@@ -1,6 +1,6 @@
 package me.daegyeo.netflixchecker.api.controller
 
-import me.daegyeo.netflixchecker.api.controller.dto.AuthLoginRequest
+import me.daegyeo.netflixchecker.api.controller.request.AuthLoginRequest
 import me.daegyeo.netflixchecker.api.service.AuthService
 import me.daegyeo.netflixchecker.shared.annotation.CheckAuth
 import org.springframework.web.bind.annotation.*
