+++ src/main/kotlin/me/daegyeo/netflixchecker/api/controller/request/AuthLoginRequest.kt
@@ -1,4 +1,4 @@
-package me.daegyeo.netflixchecker.api.controller.dto
+package me.daegyeo.netflixchecker.api.controller.request
 
 data class AuthLoginRequest(
     val email: String,
